feat(workspace): add runt-workspace crate and dev-mode notebooks directory
Create a new runt-workspace crate to centralize workspace and dev mode utilities
(is_dev_mode, get_workspace_path, worktree_hash, daemon_base_dir). This improves
the dependency graph by removing the need for notebook and other crates to depend
on runtimed just for workspace utilities.

Add default_notebooks_dir() function that returns {workspace}/notebooks in dev
mode and ~/notebooks otherwise. Update notebook save dialogs and kernel working
directory logic to use this new function, so test notebooks stay organized within
the workspace during development.

Update runtimed to re-export workspace functions for backwards compatibility.
Add notebooks/ to .gitignore so dev notebooks aren't accidentally committed.
