+++ .gitignore
@@ -11,6 +11,9 @@ target/
 .ipynb_checkpoints
 Untitled*.ipynb
 
+# Dev mode notebooks directory
+notebooks/
+
 # Local sqlite db files
 *.db
 *.db-wal
